Move old TS config stuff to legacy
